Compute user membership based on group members
The current implementation of identitysync is wasteful in terms
of API calls.

We can assume that there are less groups then users, and actually
we will sync much more groups than users using some sync filter.

Because that, a better approach would be:
 - List the groups with members
 - Then list the users
 - build the user groups based on the listed groups, avoiding additonal
 calls.

For this we need to change te order of the sync code and pass the
groups to the corresponding provider ListUsers.

Here I only changed it for okta, but we need to change it for all
even if later is ignored